2.3.20

+++ atec-cache-apcu/uninstall.php
@@ -1,5 +1,5 @@
 <?php
-defined('ABSPATH') || exit;
+if (!defined('WP_UNINSTALL_PLUGIN')) {	exit; }
 if (!defined('ATEC_LOADER')) require __DIR__ . '/includes/ATEC/LOADER.php';
 
 delete_option('atec_wpca_last_cache');
